Move lesson staleness render test
